requirements.txt
earthengine-api
flask
flask-SQLAlchemy
flask_migrate
folium
geemap
geopandas
h5py
leafmap
matplotlib
mlstac
numpy
opencv-python
pandas
pillow
pyOpenSSL
python-dotenv
rasterio
retry
fiona
scikit-image
scikit-learn==1.5.1
segmentation_models_pytorch
torch
torchmetrics
transformers
wandb
google-auth
flask_session
gunicorn
redis
geoalchemy2
psycopg2-binary
shapely